// src/shared/enums/formSteps.enum.ts
export enum FormStepsEnum {
  step_first = 'step_first',
  step_second = 'step_second',
  step_third = 'step_third',
  step_fourth = 'step_fourth',
  step_fifth = 'step_fifth',
}

export enum FormSubStepsEnum {
  sub_step_first = 'sub_step_first',
  sub_step_second = 'sub_step_second',
  sub_step_third = 'sub_step_third',
  sub_step_fourth = 'sub_step_fourth',
}